refactor(radio): separate runtime sorting and timestamp logic into utils.ts with dedicated test file
